Broke up main.rs into modules and plugins
